GH-1 Add initial datamodel and pass-through vol fitter
Add a bare-bones domain model which begins to define the input
and output API so that we may begin building around this in
subsequent commits. For now, add a placeholder pass-through
vol "fitter" which does as little work as possible, simply
returning the midpoint of the per-strike IVs. This will be removed
later, but is just to let us get a running application off the ground.

Also remove the "hello world" main and test classes that I had added
earlier; these are no longer needed now that we have some real code.

At this point, a note on the API is in order: I am choosing to
represent the fitted volatility surface simply as the fitted
vols per strike/expiry. This is a model-independent representation,
meaning it enables us to swap out the vol model without worry.
A downside is that we lose the ability to rely on the model
to interpolate the surface (whether in strike or expiry space)!

As a suggested way to mitigate this, a future extension would
be to export the first and second derivatives of the vol curve
w.r.t. strike along with the vol at each strike. Then, consumers
could interpolate in strike using cubic splines.

Finally, note that another possible future extension is to
use the flyweight pattern to reduce allocations of Option
objects. This is left out for now as a premature optimization.