Add Black formatting check to Actions workflow
We will use the Black formatting tool to enforce code
formatting in this project. Add a Black check to the
Actions workflow so that compliance with Black will be checked on each push. It will still be the responsibility of the author to run Black manually.

Leave the flake8 linter in the workflow as well, because it catches more things that just formatting (e.g., unused variables).