GH-1 Package the package so tests can be run
When using the src/ directory structure and placing tests
outside of the application code itself, we actually need
to package and install the code within the src directory
so that pytests can access it. This will also allow us to
package the package for distribution as well, but here we
just focus on making it available to pytest.

- Add pyproject.toml, which declares that we want to use
  setuptools to package the project.
- Add setup.cfg to specify the package information. This is
  analogous to setup.py, but seems to be the more modern approach.
- We tell setup.cfg to look in the src/ directory, specifically.

Then, one can run

pip install -e .

from the command line, which will allow pytest to run on the
files in the src dir.