GH-1 Add RawIVSupplier
Add AbstractRawIVSupplier, which will stub out the supplying of
our IV input data. Add a toy concrete implementation of it which
loads sample input data from a CSV file containing OptionMetrics
vended data. Add the OptionMetrics reference manual to the papers
directory to document the format of the expected data (which already
is stored in our sample input data directory).

Right now this assumes the dataframe contains only a single vol
surface, which is not the case in practice. Subsequent commits
will flesh out how this code is used.