GH-1 Add FinalIVConsumer
Add AbstractFinalIVConsumer to stub out the endpoint of the
downstream consumer(s) that will listen to the final IV
surfaces that are produced by the vol fitter.

Add a sample implementation, PickleFinalIVConsumer, which just
writes the surface to a pickle file on disc. I chose pickle
instead of JSON because it can automatically serialize my
custom dataclasses, as well as datetime objects, which JSON
can only handle with a custom serializer. A tradeoff is that
pickle is Python-specific and not human readable.