GH-1 Add service layer and CLI entrypoint
Add VolfitterService, a service layer class which orchestrates
the interactions amongst the various components of the system.
Importantly, VolfitterService depends only on abstractions, both
so that it can be unit tested and so that we can easily configure
the application to run "live" or on "toy data" or in any other mode.

Also add an endpoint which will start and run the application. This
is the main entrypoint, and it will create concrete implementations
of all the abstract base classes and wire the system together.

Add configuration to setup.cfg to create an executable shell script
which hooks into this entrypoint. Thus, after doing a pip install,
one can simply call

volfitter

at the command line and the application will run.

- For now, just do all the dependency creation/wiring directly
  in the entrypoint's run method. A subsequent comment will pull
  this out to a bootstrapper. We'll also give the user the ability
  to configure whether the application is run in "toy data" or
  "live" mode.
- For nor, just hardcode the paths to the input and output files.
  A subsequent comment will pull this out to a configuration layer.
- Also, right now, it is creating a "vol surface" from ALL days in
  January 2020. Obviously this is meaningless. Again, we'll fix this
  up later; this commit is just to wire things together and get the
  data flowing through.