GH-1 Trigger vol fits on a timer
Use the APScheduler package to trigger vol fits on a timer.
The fit interval, in seconds, is exposed as a user parameter,
VOLFITTER_FIT_INTERVAL_S, defaulting to 10s.

Use a BlockingScheduler so that the process will not exit
after starting the scheduler. The process will now run
indefinitely until it is killed.

Manually setting the local timezone is necessary to supress a
warning stemming from APScheduler's deprecated use of pytz
functionality.