GH-1 Loop through daily vol curves in input file
Make the behavior of the fitter a bit more realistic when
it is run on the sample input file of daily option data.
Specifically, every time the timer fires to trigger a new
run on the configured interval, pull in the next raw vol
surface in the input file so that we will eventually loop
through all of them. When we reach the end, loop back around
to the beginning so that the process runs indefinitely. (This
is somewhat arbitrary, but an attempt to mimic how it would
run in prod.)

As part of this, the fitter needs to think that the "current
time" is the time of the current vol surface in the input file,
not the "real" time when the application is running. To achieve
this, add a CurrentTimeSupplier which in this case just supplies
the "current times" from the same input file. This sort of thing
will also be necessary for backtesting/simulations.

- Rather than the user specifying the input data filename, specify
  a substring, defaulting to "option_data," and load all files
  in the input dir which contain this substring in their filename.
- Cache the dataframe so that it is read once from disc at startup
  and then kept in memory for subsequent runs.
- Also, add the datetime of the vol surface to RawIVSurface and
  FinalIVSurface.