GH-1 Add functional test
Add a functional test which stubs out the I/O adaptors
with test adaptors but uses the real implementations of
all other internals of VolfitterService (not much at the
moment, but that will change).

Add stock test data to conftest.py so that it can be reused
by other tests.