GH-2 Add PricingSupplier
Create and wire in PricingSupplier, though it is
so far unused.

Some pricing values (delta, gamma, vega, theta) we
read directly from the sample data CSVs. Others, such
as moneyness and time to expiry, we must calculate
ourselves.

Delta, gamma, and theta will not actually be used in
this project. Although it is therefore a violation
of the "You Ain't Gonna Need It" principle (YAGNI) to
include them, since this is a toy project, I still include
them to demonstrate the sorts of things that would live
in the Pricing class in a fuller-fledged application.