GH-2 Discard empty and one-sided markets
A more sophisticated fitter could still find a
way to use one-sided markets, for example by
trying not to fit too close to the one side that
does exist. As this is a toy vol fitter, we shall
just discard them entirely for simplicity.

Note: This check will throw away ALL options on the
day of expiry in my sample data, because that data is
taken right at the close. At the close on the day of
expiry, all the IVs (and Greeks, etc) are NaN.