GH-2 Mark expiries with not enough valid strikes as FAILed
Add a Status tag to RawIVCurve and FinalIVCurve. If a RawIVCurve
is WARN or FAIL, the fitter should propagate this and not attempt
to fit the expiry. Later on, we will also have final curve
validation that can set the final curve status.

Expose a user parameter,

VOLFITTER_RAW_IV_FILTERING_CONFIG_MIN_VALID_STRIKES_FRACTION,

which specifies the minimum number of valid strikes an expiry
must have to be fit. We also hard-code a floor of three strikes,
because we need at least this many points to fit an SVI curve.
(Three is probably an absolute minimum number of points to fit
many models, not just SVI.)