GH-4 Add final IV validation w/ crossed PnL check
This validation check calculates the total amount a curve
is through the market, in dollar terms, and compares that
to configure thresholds.

Expose parameters for both WARN and FAIL thresholds.

The result of breaching validation, in our toy application,
is simply to set the status on the curve. It is then up to
downstream users of the vol curves to decide how to handle
the failures (or warnings).

Introduce a CompositeFinalIVValidator to show how additional
validators could easily be added, analogously to the setup
for raw IV filters, although the composite is unnecessesary
at the moment when there is only one validator.