GH-6 Add visualization Jupyter notebook
Add a notebook which will run the fitter on the sample
data and visualize the out. It can visualize both single
curves and the entire surface. When plotting a curve, the
market IV quotes are also displayed.

The notebook can also run the fitter in a loop over all days
in the sample data, producing a time evolution "movie" of
the curve or surface. This is particularly interesting
because the sample data is for the year 2020, so one can
observe the effects of the covid-19 pandemic in Feb/March of
that year.