Don't include as much output in the Jupyter notebook in the repo
While the 3D plot evolution cell correctly overwrites the old
plot at each timestep, for some reason it then still displays
ALL plots when one kills the cell via a keyboard interrupt.
Not a huge problem when using this locally, but we don't
need to store all that output in the notebook in the repo.