Specify default data/log dirs relative to project root
Previously, the volfitter would fail if not run from the
project root, because the default data and log dirs
were relative to the dir from which the process was run.

Now, they are relative to the location of the config code
file, which presumably should not change. The command-line
application can now be run from anywhere, and the notebook
does not need directory defaults overwritten.

Note that the printout of the defaults in the notebook references
my personal machine's directory structure. This is an unfortunate
but harmless side-effect of the printout. When the package is
installed on any other machine, the default should be set to
the correct path for that machine automatically, so this should
be fine.