Delete requirements.txt
Requirements.txt was empty since I moved the dependencies
into setup.cfg. I had deliberately left requirements.txt even
though it was empty as a way to document the fact that it was
"still there" but not currently being used, but may as well
just get rid of it.

Do note, though, that the two files do different things and
have different intentions: setup.cfg specifies the broadest
possible range of dependencies with which the application will
work and is intended to specify what must be present when installing
and deploying the application.

By contrast, requirements.txt specificies a tigher set of
dependencies (eg, pinned versions) intended to reproduce an
exact environment for development purposes.

See https://stackoverflow.com/questions/43658870/requirements-txt-vs-setup-py

But in this project, since it's just me developing on my one
machine, I don't need a requirements.txt.